Move the scroll ID from the DELETE /_search/scroll URL to its body.
Scroll IDs can be too long for an HTTP request header, and so should
only be put in the body. Putting them in the URL was deprecated in
ElasticSearch 7.0.0 (April 2019).

The SearchClient.scroll method already does this, it was only
SearchClient.deleteScroll that did not.

See https://www.elastic.co/guide/en/elasticsearch/reference/current/clear-scroll-api.html